feat: add constants and types mod
