+++ generator/rust/src/constants.rs
@@ -1,0 +1,3 @@
+pub const MAX_DEPLOY_ARGS: usize = 10;
+pub const MAX_CONTENT_SIZE: usize = 1024;
+pub const MAX_STRING_LEN: usize = 64;
